extracted staking services repository
